Add default extra cli options by browser
This still doesn't work the way you'd like - it will still use the
existing profile if there is a Chrome running. This can be bypassed
with e.g. -user-data-dir=/tmp/profiled, but this invokes the 'first
time run' behavior, asking for permission to collect bug reports,
and opening two tabs.